pyproject.toml
[project]
name = "prj1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.2.4",
    "tavily-python>=0.7.10",
]
